updat

+++ saleProduct/ChildForm/UC_Customers.cs
@@ -30,7 +30,7 @@ namespace saleProduct.ChildForm
         public UC_Customers()
         {
             InitializeComponent();
-            conn = new SqlConnection("server=WINDOWS-11; database=SaleStoreX; integrated security=true");
+            conn = new SqlConnection("server=WINDOWS-11; database=XStore; integrated security=true");
         }
 
         private void UC_Customers_Load(object sender, EventArgs e)
@@ -47,12 +47,19 @@ namespace saleProduct.ChildForm
             dgrCustomer.DataSource = dt;
             conn.Close();
         }
-
+        private void ClearData()
+        {
+            txtCustomerID.Clear();
+            txtCustomerName.Clear();
+            txtCustomerPhone.Clear();
+            txtAddress.Clear();
+            txtCustomerName.Focus();
+        }
         private void btnAddCustomer_Click(object sender, EventArgs e)
         {
             conn.Open();
-            string insert = "INSERT INTO Customer (CustomerID, CustomerName, CustomerPhone, Address) " +
-                            "VALUES (@id, @name, @phone, @address)";
+            string insert = @"INSERT INTO Customer (CustomerID, CustomerName, CustomerPhone, Address)
+                              VALUES (@id, @name, @phone, @address)";
             SqlCommand cmd = new SqlCommand(insert, conn);
 
             cmd.Parameters.Add("@id", SqlDbType.VarChar).Value = txtCustomerID.Text;
@@ -60,10 +67,16 @@ namespace saleProduct.ChildForm
             cmd.Parameters.Add("@phone", SqlDbType.VarChar).Value = txtCustomerPhone.Text;
             cmd.Parameters.Add("@address", SqlDbType.VarChar).Value = txtAddress.Text;
 
-            cmd.ExecuteNonQuery();
+            int i = cmd.ExecuteNonQuery();
             conn.Close();
 
-            FillData();
+            if (i > 0)
+            {
+                FillData();
+                ClearData();
+                MessageBox.Show(this, "Added successfully!", "Result",
+                    MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
         }
 
         private void dgrCustomer_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -80,32 +93,54 @@ namespace saleProduct.ChildForm
 
         private void btnUpdateCustomer_Click(object sender, EventArgs e)
         {
-            conn.Open();
-            string update = "UPDATE Customer SET CustomerName=@name, CustomerPhone=@phone, Address=@address " +
-                            "WHERE CustomerID=@id";
-            SqlCommand cmd = new SqlCommand(update, conn);
+            if (MessageBox.Show(this, "Do you want to edit?", "Question",
+                MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            {
+                conn.Open();
+                string update = @"UPDATE Customer 
+                                  SET CustomerName=@name, CustomerPhone=@phone, Address=@address
+                                  WHERE CustomerID=@id";
+                SqlCommand cmd = new SqlCommand(update, conn);
 
-            cmd.Parameters.Add("@id", SqlDbType.VarChar).Value = txtCustomerID.Text;
-            cmd.Parameters.Add("@name", SqlDbType.VarChar).Value = txtCustomerName.Text;
-            cmd.Parameters.Add("@phone", SqlDbType.VarChar).Value = txtCustomerPhone.Text;
-            cmd.Parameters.Add("@address", SqlDbType.VarChar).Value = txtAddress.Text;
+                cmd.Parameters.Add("@id", SqlDbType.VarChar).Value = txtCustomerID.Text;
+                cmd.Parameters.Add("@name", SqlDbType.VarChar).Value = txtCustomerName.Text;
+                cmd.Parameters.Add("@phone", SqlDbType.VarChar).Value = txtCustomerPhone.Text;
+                cmd.Parameters.Add("@address", SqlDbType.VarChar).Value = txtAddress.Text;
 
-            cmd.ExecuteNonQuery();
-            conn.Close();
+                int i = cmd.ExecuteNonQuery();
+                conn.Close();
 
-            FillData();
+                if (i > 0)
+                {
+                    FillData();
+                    ClearData();
+                    MessageBox.Show(this, "Updated successfully!", "Result",
+                        MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+            }
         }
 
         private void btnDeleteCustomer_Click(object sender, EventArgs e)
         {
-            conn.Open();
-            string delete = "DELETE FROM Customer WHERE CustomerID=@id";
-            SqlCommand cmd = new SqlCommand(delete, conn);
-            cmd.Parameters.Add("@id", SqlDbType.VarChar).Value = txtCustomerID.Text;
-            cmd.ExecuteNonQuery();
-            conn.Close();
+            if (MessageBox.Show(this, "Do you want to delete?", "Question",
+                 MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            {
+                conn.Open();
+                string delete = "DELETE FROM Customer WHERE CustomerID=@id";
+                SqlCommand cmd = new SqlCommand(delete, conn);
+                cmd.Parameters.Add("@id", SqlDbType.VarChar).Value = txtCustomerID.Text;
 
-            FillData();
+                int i = cmd.ExecuteNonQuery();
+                conn.Close();
+
+                if (i > 0)
+                {
+                    FillData();
+                    ClearData();
+                    MessageBox.Show(this, "Deleted successfully!", "Result",
+                        MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+            }
         }
     }
